backend/requirements.txt
alembic==1.12.0
aniso8601==9.0.1
APScheduler==3.10.4
attrs==23.1.0
bcrypt==4.0.1
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
defusedxml==0.7.1
et-xmlfile==1.1.0
Flask==2.3.3
Flask-APScheduler==1.13.0
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.2
flask-marshmallow==0.15.0
Flask-Migrate==4.0.5
flask-restx==1.1.0
Flask-SQLAlchemy==3.1.1
greenlet==2.0.2
gunicorn==21.2.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
Mako==1.2.4
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
mysqlclient==2.1.0
openpyxl==3.1.2
packaging==23.1
paho-mqtt==1.6.1
PyJWT==2.8.0
pyModbusTCP==0.2.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
referencing==0.30.2
requests==2.31.0
rpds-py==0.10.3
six==1.16.0
SQLAlchemy==2.0.21
typing_extensions==4.8.0
tzlocal==5.0.1
urllib3==2.0.5
Werkzeug==2.3.7
